// client/src/redux/action-types/location.types.ts
export enum LocationTypes {
  FETCH_LOCATION = "FETCH_LOCATION",
  FETCH_LOADING = "FETCH_LOADING",
  DELETE_LOCATION = "DELETE_LOCATION",
  ADD_LOCATION = "ADD_LOCATION",
  UPDATE_LOCATION = "UPDATE_LOCATION",
  SORT_LOCATION_BY_NAME = "SORT_LOCATION_BY_NAME",
  SORT_LOCATION_BY_DATE_CREATED = "SORT_LOCATION_BY_DATE_CREATED",
  FETCH_SPECIFIC_LOCATION = "FETCH_SPECIFIC_LOCATION",
  THROW_ERROR = "THROW_ERROR",
}

export enum CommentTypes {
  COMMENT_LOADING = "COMMENT_LOADING",
  ADD_COMMENT = "ADD_COMMENT",
  GET_COMMENT = "GET_COMMENT",
  UPDATE_COMMENT = "UPDATE_COMMENT",
  DELETE_COMMENT = "DELETE_COMMENT",
  FETCH_SPECIFIC_COMMENT = "FETCH_SPECIFIC_COMMENT",
  GET_COMMENT_AUTHOR = "GET_COMMENT_AUTHOR",
}